+++ src/TechChallenge.Application/Features/Estoque/BaixarEstoque/BaixarEstoqueCommand.cs
@@ -3,5 +3,5 @@ namespace TechChallenge.Application.Features.Estoque.BaixarEstoque;
 public class BaixarEstoqueCommand
 {
     public Guid ProdutoId { get; set; }
-    public double Quantidade { get; set; }
+    public int Quantidade { get; set; }
 }
